refac(*): Remove olb code. Move introspection
